+++ test_external_plugins/snowpark_hello_single_command/src/snowflakecli/test_plugins/snowpark_hello/plugin_spec.py
@@ -26,5 +26,5 @@ def command_spec():
     return CommandSpec(
         parent_command_path=CommandPath(["snowpark"]),
         command_type=CommandType.SINGLE_COMMAND,
-        typer_instance=commands.app,
+        typer_instance=commands.app.create_instance(),
     )