+++ tests_e2e/test_installation.py
@@ -12,21 +12,22 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import subprocess
 from pathlib import Path
 
 import pytest
 
+from tests_e2e.conftest import subprocess_check_output
+
 
 @pytest.mark.e2e
 def test_snow_help(snowcli, snapshot):
-    output = subprocess.check_output([snowcli, "--help"], encoding="utf-8")
+    output = subprocess_check_output([snowcli, "--help"])
     snapshot.assert_match(output)
 
 
 @pytest.mark.e2e
 def test_snow_sql(snowcli, test_root_path, snapshot):
-    output = subprocess.check_output(
+    output = subprocess_check_output(
         [
             snowcli,
             "--config-file",
@@ -37,16 +38,13 @@ def test_snow_sql(snowcli, test_root_path, snapshot):
             "-c",
             "integration",
         ],
-        encoding="utf-8",
     )
     snapshot.assert_match(output)
 
 
 @pytest.mark.e2e
 def test_snow_streamlit_init(temp_dir, snowcli, snapshot):
-    output = subprocess.check_output(
-        [snowcli, "streamlit", "init", "streamlit_test"], encoding="utf-8"
-    )
+    output = subprocess_check_output([snowcli, "streamlit", "init", "streamlit_test"])
     snapshot.assert_match(output)
     expected_generated_file = f"{temp_dir}/streamlit_test/pages/my_page.py"
     assert Path(
@@ -56,7 +54,7 @@ def test_snow_streamlit_init(temp_dir, snowcli, snapshot):
 
 @pytest.mark.e2e
 def test_command_from_external_plugin(snowcli, test_root_path, snapshot):
-    output = subprocess.check_output(
+    output = subprocess_check_output(
         [
             snowcli,
             "--config-file",
@@ -67,6 +65,5 @@ def test_command_from_external_plugin(snowcli, test_root_path, snapshot):
             "-c",
             "integration",
         ],
-        encoding="utf-8",
     )
     snapshot.assert_match(output)